No longer use RAM to hold arr in memory
